+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Project logs are now overwritten on each deployment so there is only one log file for each project. This will conserve space as well as reduce complexity of the project when retrieving pipelines via the API
 * Overhauled the response of the `pipelines` endpoint to include a `id`, `updated_at`, and `status` key for each record
 * Fixed various path bugs that wouldn't resolve properly on Windows (used `os.path.join()`)
+* Changed all datetime/timestamp fields to use `utc` time
 
 ## v0.15.0 (2021-11-13)